// apps/client/app/sync/apiTypes.ts
export const ApiMessageSchema = z.object({
    id: z.string(),
    seq: z.number(),
    localId: z.string().nullish(),
    content: z.discriminatedUnion('t', [
        z.object({
            t: z.literal('encrypted'),
            c: z.string(), // Base64 encoded encrypted content
        }),
        z.object({
            t: z.literal('plaintext'),
            c: z.string(), // Plain JSON string
        }),
    ]),
    createdAt: z.number(),
});

export type ApiMessage = z.infer<typeof ApiMessageSchema>;